README.md
In Alien Invasion, the player controls a ship that appears at the bottom center of the screen. The player can move the ship right and left, using the arrow keys and shoot missiles using the spacebar. When the game begins, a fleet of aliens fills the sky and moves across and down the screen. The player shoots and destroys the aliens. If the player shoots all the aliens, a new fleet appears that moves faster than the previous fleet. If any alien hits the player's ship or reaches the bottom of the screen, the player loses a ship. If a player loses three ships, the game ends.